pyproject.toml
[project]
name = "kitaev-spinhalf-electirc-effect"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.15.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "physics-tenpy>=1.1.0",
    "pygame>=2.6.1",
]
